remove preset and use local-preset for dev
