--- preset.js
@@ -1,12 +1,0 @@
-function previewAnnotations(entry = []) {
-  return [...entry, require.resolve("./dist/esm/preset/preview")];
-}
-
-function managerEntries(entry = []) {
-  return [...entry, require.resolve("./dist/esm/preset/manager")];
-}
-
-module.exports = {
-  managerEntries,
-  previewAnnotations,
-};
